Move command processing to separate files
