+++ utils/profile_tool/__init__.py
@@ -1,0 +1,2 @@
+from .sub import command_sub
+from .stats import command_stats
